+++ packages/cloud-cognitive/src/components/NotificationsPanel/NotificationsPanel.stories.js
@@ -103,12 +103,9 @@ const Template = (args) => {
   useEffect(() => {
     let unreadTimer;
     if (open && hasUnreadNotifications) {
-      const tempData = JSON.parse(JSON.stringify(notificationsData));
+      const tempData = [...notificationsData];
       tempData.forEach((element) => {
         element.unread = false;
-        // convert timestamp back to date object, otherwise this is
-        // a UTC date string and the component is expecting a date object
-        element.timestamp = new Date(element.timestamp);
       });
       unreadTimer = setTimeout(() => {
         setHasUnreadNotifications(false);