+++ packages/cloud-cognitive/CHANGELOG.md
@@ -3,6 +3,91 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.88.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.87.3...@carbon/ibm-cloud-cognitive@0.88.0) (2021-09-27)
+
+
+### Features
+
+* **EditSidePanel:** [#1266](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1266) edit side panel ([#1280](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1280)) ([25f2b6e](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/25f2b6ebcbdb78304caa16ce4b2ece56ecc83ea5))
+
+
+
+
+
+## [0.87.3](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.87.2...@carbon/ibm-cloud-cognitive@0.87.3) (2021-09-24)
+
+
+### Bug Fixes
+
+* **EmptyStates:** give svg elements unique ids to avoid display issues ([#1290](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1290)) ([72deea8](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/72deea83d493930ba03ee938c4aa1f588662407f))
+
+
+
+
+
+## [0.87.2](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.87.1...@carbon/ibm-cloud-cognitive@0.87.2) (2021-09-23)
+
+
+### Bug Fixes
+
+* **SidePanel:** styles refactor/fix ([#1285](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1285)) ([bf78f15](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/bf78f157192cefdf0e0977895cf144f22d9012ab))
+
+
+
+
+
+## [0.87.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.87.0...@carbon/ibm-cloud-cognitive@0.87.1) (2021-09-23)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+# [0.87.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.86.1...@carbon/ibm-cloud-cognitive@0.87.0) (2021-09-23)
+
+
+### Features
+
+* **SidePanel:** add prop to prevent closing panel on click outside ([#1288](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1288)) ([8dfa143](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/8dfa143ac88442e1a770d217298c9c95a07fe590))
+
+
+
+
+
+## [0.86.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.86.0...@carbon/ibm-cloud-cognitive@0.86.1) (2021-09-23)
+
+
+### Bug Fixes
+
+* storybook setting masking missing component color setting ([#1241](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1241)) ([38c01ec](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/38c01ec0fa1992b358aced18b714d77eda52bf09))
+
+
+
+
+
+# [0.86.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.85.2...@carbon/ibm-cloud-cognitive@0.86.0) (2021-09-23)
+
+
+### Features
+
+* update Carbon versions and dependencies to latest ([#1282](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1282)) ([b1451cf](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/b1451cf5c91b75c1bd4e55ae2bdb47f025f33163))
+
+
+
+
+
+## [0.85.2](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.85.1...@carbon/ibm-cloud-cognitive@0.85.2) (2021-09-22)
+
+
+### Bug Fixes
+
+* keep page header above content on scroll ([#1281](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1281)) ([f2d95e4](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/f2d95e434f37945d6d249631f44f8ce4789b7a58))
+
+
+
+
+
 ## [0.85.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.85.0...@carbon/ibm-cloud-cognitive@0.85.1) (2021-09-22)